requirements.txt
python-docx==0.8.11
pyyaml>=6.0